sql update  + format pe inventar + inventar intra gata

+++ src/main/java/projectManager/repository/dao/Cod3DAO.java
@@ -11,5 +11,5 @@ public interface Cod3DAO extends GenericDAO<Cod3, Integer> {
 
     Cod3 findByBarcode(String code);
 
-    void setStare(byte stare, int id);
+    void setStare(int stare, int id);
 }
